TripAI/tripai/requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.1.1
groq==0.4.0
python-dotenv==1.0.0
SQLAlchemy==2.0.21
Werkzeug==2.3.7
forex-python==1.8
fpdf2==2.7.5
qrcode==7.4.2